set button type to button to prevent page reload on edge
