+++ projects/media-viewer/src/lib/annotations/comment-set/comment/comment.component.html
@@ -33,25 +33,28 @@
   <ng-container *ngIf="selected || this.editable">
     <div class="aui-comment__footer commentBtns">
       <button *ngIf="!editable; else Save"
-              class="govuk-button" type="submit" role="button"
+              class="govuk-button" type="button" role="button"
               (click)="onEdit()">
         Edit
       </button>
       <button *ngIf="!editable; else Cancel"
-              (click)="onDelete()"
+              type="button" role="button"
               class="govuk-button govuk-button--secondary"
-              role="button">
+              (click)="onDelete()">
         Delete
       </button>
     </div>
   </ng-container>
   <ng-template #Save>
-    <button class="govuk-button" type="submit" role="button" (click)="onSave()">
+    <button class="govuk-button" type="button" role="button"
+            (click)="onSave()">
       Save
     </button>
   </ng-template>
   <ng-template #Cancel>
-    <button class="govuk-button govuk-button--secondary" role="button" (click)="onCancel()">
+    <button type="button" role="button"
+            class="govuk-button govuk-button--secondary"
+            (click)="onCancel()">
       Cancel
     </button>
   </ng-template>
